LlamaIndex_PGI/pyproject.toml
[project]
name = "llamaindex-pgi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "llama-index>=0.14.13",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-graph-stores-neo4j>=0.5.1",
    "llama-index-llms-anthropic>=0.10.6",
    "llama-index-llms-openai>=0.6.15",
    "python-dotenv>=1.2.1",
]
